+++ resolvers/Mutation.js
@@ -1,6 +1,8 @@
 import {pubsub} from "./context.js"
 import { orderList, addToOrderList } from "../controllers/order.js"
-import {dbMutation} from "../db/connection.js"
+import {dbMutation, dbQuery} from "../db/connection.js"
+import {queryAllItem} from "../db/utility.js"
+
 // ! passing context from Appolo server constructor is not working
 const Mutation = {
     createOrder(parent, {order}, {}, info){
@@ -15,11 +17,10 @@ const Mutation = {
     async createItem(parent, {data}, {}, info){
         console.log("recevied data:", data)
         let itemId = "item"+Math.floor(Math.random()*1000)
-        console.log(typeof(itemId))
         let result = await dbMutation(`INSERT INTO \`Item\` VALUES('${itemId}', '${data.img}', ${data.price} )`)
         //! schema input type is not well defined, so we cannot insert full info into db
         result = await dbMutation(`INSERT INTO \`Item_Trans\` VALUES('${Math.floor(Math.random()*1000)}', '${data.name}' ,'zh', '', '', '${itemId}')`)
-        console.log(result)
+
         return data
     },
     updateItem(parent, {id, data}, {}, info){
@@ -27,9 +28,12 @@ const Mutation = {
         console.log("recevied data:", data)
         // todo
     },
-    deleteItem(parent, {id, data}, {}, info){
-        console.log("id:", id)
-        // todo
+    async deleteItem(parent, {id}, {}, info){
+        let allItem = await queryAllItem()
+        let [result] = allItem.filter(e=>e.itemId===id)
+        console.log("allItem", allItem)
+        await dbMutation(`DELETE FROM \`Item\` WHERE \`id\`='${id}'`)
+        return result
     }
 }
