+++ resolvers/utility.js
@@ -1,3 +1,4 @@
+
 const getOrderById = (id, orderList) => {
     let order = orderList.filter((e)=>e.id===id)
     return order 
@@ -60,5 +61,5 @@ export {getOrderById,
     prepareItem, 
     preapreOrderItem, 
     prepareOrderIdList,
-    prepareOrderList
+    prepareOrderList,
 }
